+++ docs/.vitepress/config.js
@@ -118,7 +118,8 @@ export default defineConfig({
         {
           text: 'Changelog',
           items: [
-            { text: 'v1.8.1 (Latest)', link: '/changelog#v1-8-1' },
+            { text: 'v1.8.2 (Latest)', link: '/changelog#v1-8-2' },
+            { text: 'v1.8.1', link: '/changelog#v1-8-1' },
             { text: 'v1.8.0', link: '/changelog#v1-8-0' },
             { text: 'v1.7.4', link: '/changelog#v1-7-4' },
             { text: 'v1.7.3', link: '/changelog#v1-7-3' },